+++ Runtime/QuickSave.Builder.cs
@@ -1,10 +1,9 @@
-﻿namespace Achieve.QuickSave
+namespace Achieve.QuickSave
 {
     public sealed partial class QuickSave<T>
     {
         public class Builder
         {
-
 #if USE_ENCRYPT
             private bool _encrypt = false;
             private int _version;
@@ -13,8 +12,6 @@
             /// <summary>
             /// 암/복호화를 사용합니다.
             /// </summary>
-            /// <param name="encryptionKey"></param>
-            /// <returns></returns>
             public Builder UseEncryption(string encryptionKey)
             {
                 _encrypt = true;
@@ -25,8 +22,6 @@
             /// <summary>
             /// Save 파일의 버전을 명명합니다. 입력하지 않으면 0을 바라봅니다.
             /// </summary>
-            /// <param name="version"></param>
-            /// <returns></returns>
             public Builder UseVersion(int version)
             {
                 _version = version;
@@ -38,9 +33,10 @@
                 var instance = new QuickSave<T>();
 #if USE_ENCRYPT
                 instance._isEncrypt = _encrypt;
-                instance._version = 0;
+                instance._version = _version;
                 instance._encryptionKey = _encryptionKey;
 #endif
+                instance.Initialize();
                 return instance;
             }
         }